+++ packages/astro/src/core/config/config.ts
@@ -1,20 +1,16 @@
 import type { Arguments as Flags } from 'yargs-parser';
 import type { AstroConfig, AstroUserConfig, CLIFlags } from '../../@types/astro';
 
-import load, { ProloadError, resolve } from '@proload/core';
-import loadTypeScript from '@proload/plugin-tsm';
 import fs from 'fs';
 import * as colors from 'kleur/colors';
 import path from 'path';
 import { fileURLToPath, pathToFileURL } from 'url';
-import * as vite from 'vite';
 import { mergeConfig as mergeViteConfig } from 'vite';
 import { AstroError, AstroErrorData } from '../errors/index.js';
 import { LogOptions } from '../logger/core.js';
 import { arraify, isObject, isURL } from '../util.js';
 import { createRelativeSchema } from './schema.js';
-
-load.use([loadTypeScript]);
+import { loadConfigWithVite } from './vite-load.js';
 
 export const LEGACY_ASTRO_CONFIG_KEYS = new Set([
 	'projectRoot',
@@ -153,7 +149,7 @@ interface LoadConfigOptions {
  * instead of the resolved config
  */
 export async function resolveConfigPath(
-	configOptions: Pick<LoadConfigOptions, 'cwd' | 'flags'>
+	configOptions: Pick<LoadConfigOptions, 'cwd' | 'flags'> & { fs: typeof fs }
 ): Promise<string | undefined> {
 	const root = resolveRoot(configOptions.cwd);
 	const flags = resolveFlags(configOptions.flags || {});
@@ -167,14 +163,14 @@ export async function resolveConfigPath(
 	// Resolve config file path using Proload
 	// If `userConfigPath` is `undefined`, Proload will search for `astro.config.[cm]?[jt]s`
 	try {
-		const configPath = await resolve('astro', {
-			mustExist: !!userConfigPath,
-			cwd: root,
-			filePath: userConfigPath,
+		const config = await loadConfigWithVite({
+			configPath: userConfigPath,
+			root,
+			fs: configOptions.fs,
 		});
-		return configPath;
+		return config.filePath;
 	} catch (e) {
-		if (e instanceof ProloadError && flags.config) {
+		if (flags.config) {
 			throw new AstroError({
 				...AstroErrorData.ConfigNotFound,
 				message: AstroErrorData.ConfigNotFound.message(flags.config),
@@ -197,7 +193,7 @@ export async function openConfig(configOptions: LoadConfigOptions): Promise<Open
 	const flags = resolveFlags(configOptions.flags || {});
 	let userConfig: AstroUserConfig = {};
 
-	const config = await tryLoadConfig(configOptions, flags, root);
+	const config = await tryLoadConfig(configOptions, root);
 	if (config) {
 		userConfig = config.value;
 	}
@@ -218,7 +214,6 @@ interface TryLoadConfigResult {
 
 async function tryLoadConfig(
 	configOptions: LoadConfigOptions,
-	flags: CLIFlags,
 	root: string
 ): Promise<TryLoadConfigResult | undefined> {
 	const fsMod = configOptions.fsMod ?? fs;
@@ -227,6 +222,7 @@ async function tryLoadConfig(
 		let configPath = await resolveConfigPath({
 			cwd: configOptions.cwd,
 			flags: configOptions.flags,
+			fs: fsMod,
 		});
 		if (!configPath) return undefined;
 		if (configOptions.isRestart) {
@@ -249,51 +245,13 @@ async function tryLoadConfig(
 			configPath = tempConfigPath;
 		}
 
-		const config = await load('astro', {
-			mustExist: !!configPath,
-			cwd: root,
-			filePath: configPath,
+		// Create a vite server to load the config
+		const config = await loadConfigWithVite({
+			configPath,
+			fs: fsMod,
+			root,
 		});
-
 		return config as TryLoadConfigResult;
-	} catch (e) {
-		if (e instanceof ProloadError && flags.config) {
-			throw new AstroError({
-				...AstroErrorData.ConfigNotFound,
-				message: AstroErrorData.ConfigNotFound.message(flags.config),
-			});
-		}
-
-		const configPath = await resolveConfigPath(configOptions);
-		if (!configPath) {
-			throw e;
-		}
-
-		// Fallback to use Vite DevServer
-		const viteServer = await vite.createServer({
-			server: { middlewareMode: true, hmr: false },
-			optimizeDeps: { entries: [] },
-			clearScreen: false,
-			appType: 'custom',
-			// NOTE: Vite doesn't externalize linked packages by default. During testing locally,
-			// these dependencies trip up Vite's dev SSR transform. In the future, we should
-			// avoid `vite.createServer` and use `loadConfigFromFile` instead.
-			ssr: {
-				external: ['@astrojs/mdx', '@astrojs/react'],
-			},
-		});
-		try {
-			const mod = await viteServer.ssrLoadModule(configPath);
-
-			if (mod?.default) {
-				return {
-					value: mod.default,
-					filePath: configPath,
-				};
-			}
-		} finally {
-			await viteServer.close();
-		}
 	} finally {
 		await finallyCleanup();
 	}
@@ -308,7 +266,7 @@ export async function loadConfig(configOptions: LoadConfigOptions): Promise<Astr
 	const flags = resolveFlags(configOptions.flags || {});
 	let userConfig: AstroUserConfig = {};
 
-	const config = await tryLoadConfig(configOptions, flags, root);
+	const config = await tryLoadConfig(configOptions, root);
 	if (config) {
 		userConfig = config.value;
 	}