+++ packages/astro/test/units/dev/restart.test.js
@@ -1,5 +1,6 @@
 import { expect } from 'chai';
 import * as cheerio from 'cheerio';
+import { fileURLToPath } from 'url';
 
 import {
 	createContainerWithAutomaticRestart,
@@ -7,6 +8,8 @@ import {
 	startContainer,
 } from '../../../dist/core/dev/index.js';
 import { createFs, createRequestAndResponse, triggerFSEvent } from '../test-utils.js';
+import { createSettings, openConfig } from '../../../dist/core/config/index.js';
+import { defaultLogging } from '../../test-utils.js';
 
 const root = new URL('../../fixtures/alias/', import.meta.url);
 
@@ -105,6 +108,74 @@ describe('dev container restarts', () => {
 			await restartComplete;
 
 			expect(isStarted(restart.container)).to.equal(true);
+		} finally {
+			await restart.container.close();
+		}
+	});
+
+	it('Is able to restart project using Tailwind + astro.config.ts', async () => {
+		const troot = new URL('../../fixtures/tailwindcss-ts/', import.meta.url);
+		const fs = createFs(
+			{
+				'/src/pages/index.astro': ``,
+				'/astro.config.ts': ``,
+			},
+			troot
+		);
+
+		const { astroConfig } = await openConfig({
+			cwd: troot,
+			flags: {},
+			cmd: 'dev',
+			logging: defaultLogging,
+		});
+		const settings = createSettings(astroConfig);
+
+		let restart = await createContainerWithAutomaticRestart({
+			params: { fs, root, settings },
+		});
+		await startContainer(restart.container);
+		expect(isStarted(restart.container)).to.equal(true);
+
+		try {
+			// Trigger a change
+			let restartComplete = restart.restarted();
+			triggerFSEvent(restart.container, fs, '/astro.config.ts', 'change');
+			await restartComplete;
+
+			expect(isStarted(restart.container)).to.equal(true);
+		} finally {
+			await restart.container.close();
+		}
+	});
+
+	it('Is able to restart project on package.json changes', async () => {
+		const fs = createFs(
+			{
+				'/src/pages/index.astro': ``,
+			},
+			root
+		);
+
+		const { astroConfig } = await openConfig({
+			cwd: root,
+			flags: {},
+			cmd: 'dev',
+			logging: defaultLogging,
+		});
+		const settings = createSettings(astroConfig, fileURLToPath(root));
+
+		let restart = await createContainerWithAutomaticRestart({
+			params: { fs, root, settings },
+		});
+		await startContainer(restart.container);
+		expect(isStarted(restart.container)).to.equal(true);
+
+		try {
+			let restartComplete = restart.restarted();
+			fs.writeFileSync('/package.json', `{}`);
+			triggerFSEvent(restart.container, fs, '/package.json', 'change');
+			await restartComplete;
 		} finally {
 			await restart.container.close();
 		}